chore: change api messages

+++ services/authentication-api/src/modules/users/user.service.ts
@@ -1,9 +1,9 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { apiErrors } from '@packages/common-resources';
+import { apiMessage } from '@packages/common-resources';
 import { Repository } from 'typeorm';
 
-import { ErrorDto } from '../common/dto/errorResponse';
+import { ErrorDto, SuccessDto } from '../common/dto/errorResponse';
 import { CreateUserDto, GetUserResponseDto, PatchUserDto, UpdateUserDto } from './dto/user.dto';
 import { UserEntity } from './entity/user.entity';
 
@@ -13,7 +13,7 @@ interface IUserService {
   createUser(user: CreateUserDto): Promise<GetUserResponseDto | ErrorDto>;
   updateUser(id: number, user: UpdateUserDto): Promise<GetUserResponseDto | ErrorDto>;
   patchUser(id: number, user: PatchUserDto): Promise<GetUserResponseDto | ErrorDto>;
-  deleteUser(id: number): Promise<void | ErrorDto>;
+  deleteUser(id: number): Promise<SuccessDto | ErrorDto>;
 }
 
 @Injectable()
@@ -29,18 +29,18 @@ export class UserService implements IUserService {
       return dbUsers.map(user => GetUserResponseDto.create(user));
     } catch (error) {
       const details = error instanceof Error ? error.message : error;
-      return ErrorDto.create({ details, ...apiErrors.INTERNAL_SERVER_ERROR });
+      return ErrorDto.create({ details, ...apiMessage.INTERNAL_SERVER_ERROR });
     }
   }
 
   async findOne(id: number): Promise<GetUserResponseDto | ErrorDto> {
     try {
       const user = await this.userRepository.findOneBy({ id });
-      if (!user) return ErrorDto.create({ ...apiErrors.USER_NOT_FOUND });
+      if (!user) return ErrorDto.create({ ...apiMessage.USER_NOT_FOUND });
       return GetUserResponseDto.create(user);
     } catch (error) {
       const details = error instanceof Error ? error.message : error;
-      return ErrorDto.create({ details, ...apiErrors.INTERNAL_SERVER_ERROR });
+      return ErrorDto.create({ details, ...apiMessage.INTERNAL_SERVER_ERROR });
     }
   }
 
@@ -51,36 +51,44 @@ export class UserService implements IUserService {
       return GetUserResponseDto.create(savedUser);
     } catch (error) {
       const details = error instanceof Error ? error.message : error;
-      return ErrorDto.create({ details, ...apiErrors.INTERNAL_SERVER_ERROR });
+      return ErrorDto.create({ details, ...apiMessage.INTERNAL_SERVER_ERROR });
     }
   }
 
-  async updateUser(id: number, user: CreateUserDto): Promise<GetUserResponseDto | ErrorDto> {
+  async updateUser(id: number, user: UpdateUserDto): Promise<GetUserResponseDto | ErrorDto> {
     try {
       await this.userRepository.update(id, user);
       return this.findOne(id);
     } catch (error) {
       const details = error instanceof Error ? error.message : error;
-      return ErrorDto.create({ details, ...apiErrors.INTERNAL_SERVER_ERROR });
+      return ErrorDto.create({ details, ...apiMessage.INTERNAL_SERVER_ERROR });
     }
   }
 
-  async patchUser(id: number, user: CreateUserDto): Promise<GetUserResponseDto | ErrorDto> {
+  async patchUser(id: number, user: PatchUserDto): Promise<GetUserResponseDto | ErrorDto> {
     try {
-      await this.userRepository.update(id, user);
+      const dbUser = await this.userRepository.findOneBy({ id });
+      if (!dbUser) return ErrorDto.create({ ...apiMessage.USER_NOT_FOUND });
+
+      const updatedUser = { ...dbUser, ...user };
+      await this.userRepository.save(updatedUser);
       return this.findOne(id);
     } catch (error) {
       const details = error instanceof Error ? error.message : error;
-      return ErrorDto.create({ details, ...apiErrors.INTERNAL_SERVER_ERROR });
+      return ErrorDto.create({ details, ...apiMessage.INTERNAL_SERVER_ERROR });
     }
   }
 
-  async deleteUser(id: number): Promise<void | ErrorDto> {
+  async deleteUser(id: number): Promise<SuccessDto | ErrorDto> {
     try {
+      const dbUser = await this.userRepository.findOneBy({ id });
+      if (!dbUser) return ErrorDto.create({ ...apiMessage.USER_NOT_FOUND });
+
       await this.userRepository.delete(id);
+      return SuccessDto.create({ ...apiMessage.DELETED_SUCCESSFULLY });
     } catch (error) {
       const details = error instanceof Error ? error.message : error;
-      return ErrorDto.create({ details, ...apiErrors.INTERNAL_SERVER_ERROR });
+      return ErrorDto.create({ details, ...apiMessage.INTERNAL_SERVER_ERROR });
     }
   }
 }
